* Filter class now can be invoked
