+++ src/Diff.php
@@ -4,11 +4,23 @@ declare(strict_types=1);
 
 namespace Whotrades\PHPMDDiff;
 
+use SebastianBergmann\Diff\Line;
+use SebastianBergmann\Diff\Parser;
 use Whotrades\PHPMDDiff\Exception\DiffException;
-use Whotrades\PHPMDDiff\Filter\XMLFilter;
+use Whotrades\PHPMDDiff\Filter\FilterXMLReportByChangelist;
 
 final class Diff
 {
+    /**
+     * @param string $phpmdFile
+     * @param string $patchFile
+     * @param string $pathPrefix
+     * @param string $filterType
+     *
+     * @return string
+     *
+     * @throws DiffException
+     */
     public function execute(string $phpmdFile, string $patchFile, string $pathPrefix, string $filterType): string
     {
         $patchRaw = @file_get_contents($patchFile);
@@ -17,6 +29,8 @@ final class Diff
             throw new DiffException("Can't load patch file.", DiffException::ERR_LOAD_FILE);
         }
 
+        $changes = $this->getChangesFromPatch($patchRaw, $pathPrefix);
+
         $reportRaw = file_get_contents($phpmdFile);
         if (empty($reportRaw)) {
             throw new DiffException("Can't load PHPMD report.", DiffException::ERR_LOAD_FILE);
@@ -24,8 +38,7 @@ final class Diff
 
         switch ($filterType) {
             case 'xml':
-                $filter = new XMLFilter();
-                break;
+                return (new FilterXMLReportByChangelist())($reportRaw, $changes);
             default:
                 if (!class_exists($filterType, true)) {
                     throw new DiffException(sprintf("Custom filter class '%s' not found", $filterType), DiffException::ERR_CUSTOM_FILTER);
@@ -35,10 +48,48 @@ final class Diff
                     throw new DiffException(sprintf("Custom filter class should extend '%s'", AbstractFilter::class), DiffException::ERR_CUSTOM_FILTER);
                 }
 
-                $filter = new $filterType();
+                return (new $filterType())($reportRaw, $changes);
         }
-
-        return $filter->execute($patchRaw, $reportRaw, $pathPrefix);
     }
 
+    /**
+     * @param string $patchRaw
+     * @param string $pathPrefix
+     *
+     * @return array
+     *
+     * @throws DiffException
+     */
+    protected function getChangesFromPatch(string $patchRaw, string $pathPrefix): array
+    {
+        if (empty($patch = (new Parser())->parse($patchRaw))) {
+            throw new DiffException("Can't load patch file.", DiffException::ERR_LOAD_DIFF);
+        }
+
+        if (substr($pathPrefix, -1, 1) !== DIRECTORY_SEPARATOR) {
+            $pathPrefix .= DIRECTORY_SEPARATOR;
+        }
+
+        $changes = [];
+        foreach ($patch as $diff) {
+            $file = $pathPrefix . substr($diff->getTo(), 2);
+            $changes[$file] = [];
+
+            foreach ($diff->getChunks() as $chunk) {
+                $lineNo = $chunk->getEnd();
+
+                foreach ($chunk->getLines() as $line) {
+                    if ($line->getType() === Line::ADDED) {
+                        $changes[$file][] = $lineNo;
+                    }
+
+                    if ($line->getType() !== Line::REMOVED) {
+                        ++$lineNo;
+                    }
+                }
+            }
+        }
+
+        return $changes;
+    }
 }
